First pass recorder and printing programs.
The 'recorder' saves data into a local SQLite database, so the user always
owns the data.

'print' will lock and open the database then gather data and print to
the terminal. This currently has issues of database locking, so if the
recorder and 'print' try to run at the same time, we can get a panic.
Simply re-running the 'print' will likely not conflict. This data-race
will need to be fixed later.