Split the various crates into smaller reusable crates.
Crates with '-lib' suffix has functionality that is specific
but can be reused.

Crates with the '-bin' suffix is are compiled into binaries that are
run by the user. The binaries take command line arguments and are
user facing.

Replace 'sqlite' crate with 'rusqlite', because it seems better supported
and has better support for an encrypted database, if we need to do that
in the future.